pyproject.toml
[project]
name = "satcomp"
version = "0.1.0"
description = "Local SAT solver competition platform"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "SATCOMP" }]
dependencies = [
  "typer>=0.9.0",
  "pyyaml>=6.0",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "jinja2>=3.1",
  "plotly>=5.18",
  "psutil>=5.9"
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[project.scripts]
satcomp = "satcomp.cli:app"

[tool.setuptools.package-data]
satcomp = ["templates/*.html", "templates/*.md", "static/*"]

[tool.setuptools.packages.find]
where = ["."]
include = ["satcomp*"]
exclude = ["data*", "runs*", "solvers*", "reports*", "tests*", "benchmarks*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"